requirements.txt
redisgraph~=2.4.4
redis~=3.5.3
requests~=2.31.0
huggingface_hub
flask~=3.0.2
openai~=1.14.1
PyYAML~=6.0.1